services/monitoring/requirements.txt
streamlit==1.28.1
plotly==5.17.0
pandas==2.1.4
numpy==1.24.4
redis==5.0.1
psycopg2-binary==2.9.9
requests==2.31.0
mlflow==2.8.1
boto3==1.34.0
python-multipart==0.0.6
aiofiles==23.2.1
asyncio==3.4.3
